Fix ThreatMap.java and updated for onlyAttack and includeDefend
Updated to check if the moving piece is at the point we are clearing. The previous changes for onlyAttacks and includeDefend was not necessary for this fix but are not included.

Why? Non-attacking moves are not threats. Defends threaten a piece if it captured. An threat on an undefended piece has value.