Update Log to no longer support regenerating log fully
There is some complexity and conflicts with board when it comes to regenerating the log. This cannot work bottom up if the board is at the top, latest state. This does not know move's follow-up action from just a string, so that is also lost and will fail building top-down.